Update demo and server tests
