+++ packages/i18n/dev/es.ts
@@ -1,0 +1,9 @@
+import type { I18nDict } from "./index.jsx";
+
+export const dict: I18nDict = {
+  hello: "hola {{ name }}, como usted?" as any,
+  goodbye: ({ name }: { name: string }) => `adios ${name}`,
+  food: {
+    meat: "carne",
+  },
+};
