+++ packages/i18n/package.json
@@ -1,10 +1,12 @@
 {
   "name": "@solid-primitives/i18n",
   "version": "1.4.1",
-  "description": "Primitive to create and use i18n primitives.",
+  "description": "Library of primitives for providing internationalization support.",
   "author": "Alexandre Mouton-Brady <amoutonbrady@gmail.com>",
   "contributors": [
-    "Tanner Scadden <tanner@scaddenfamily.com>"
+    "Tanner Scadden <tanner@scaddenfamily.com>",
+    "YISH <mokeyish@hotmail.com>",
+    "Damian Tarnawski <gthetarnav@gmail.com>"
   ],
   "license": "MIT",
   "homepage": "https://github.com/solidjs-community/solid-primitives/tree/main/packages/i18n",
@@ -16,8 +18,10 @@
     "name": "i18n",
     "stage": 3,
     "list": [
-      "createI18nContext",
-      "createChainedI18n"
+      "flatten",
+      "resolveTemplate",
+      "translator",
+      "chained"
     ],
     "category": "Utilities"
   },
@@ -42,7 +46,6 @@
   "scripts": {
     "dev": "tsx ../../scripts/dev.ts",
     "build": "tsx ../../scripts/build.ts",
-    "vitest": "vitest -c ../../configs/vitest.config.ts",
     "test": "pnpm run vitest",
     "test:ssr": "pnpm run vitest --mode ssr"
   },
@@ -57,7 +60,6 @@
     "solid-js": "^1.6.12"
   },
   "dependencies": {
-    "@solid-primitives/context": "workspace:^",
     "@solid-primitives/utils": "workspace:^"
   },
   "devDependencies": {
