+++ packages/i18n/test/server.test.ts
@@ -1,84 +1,73 @@
-import { createSignal } from "solid-js";
-import { describe, expect, it } from "vitest";
-import {
-  createChainedI18nContext,
-  createChainedI18nDictionary,
-  createI18nContext,
-} from "../src/index.js";
-import { dict } from "./setup.js";
+import { describe, expect, test } from "vitest";
+import * as i18n from "../src/new-i18n.js";
+import { en_dict } from "./setup.js";
 
-describe("createI18nContext", () => {
-  it("test locale switching", async () => {
-    const [t, { add, locale }] = createI18nContext(dict, "en");
-    Object.entries(dict).forEach(([lang, translations]) => add(lang, translations));
-    locale("en");
-    expect(t("hello", { name: "Tester" })).toBe("hello Tester, how are you?");
-    locale("fr");
-    expect(t("hello", { name: "Tester" })).toBe("bonjour Tester, comment vas-tu ?");
+describe("template", () => {
+  test("identity template resolver", () => {
+    expect(i18n.identityResolveTemplate("hello!")).toBe("hello!");
+
+    expect(i18n.identityResolveTemplate("hello {{name}}!", { name: "Tester" })).toBe(
+      "hello {{name}}!",
+    );
+  });
+
+  test("default template resolver", () => {
+    expect(i18n.resolveTemplate("hello!")).toBe("hello!");
+
+    expect(i18n.resolveTemplate("hello {{name}}!", { name: "Tester" })).toBe("hello Tester!");
   });
 });
 
-describe("createChainedI18nContext", () => {
-  it("Context should be null if setContext !== true", async () => {
-    const [, useI18nContext] = createChainedI18nContext(
-      { dictionaries: dict, locale: "en" },
-      false,
-    );
+describe("flatDict", () => {
+  test("flatDict", () => {
+    const flat = i18n.flatten(en_dict);
 
-    const context = useI18nContext();
-
-    expect(context).toBe(null);
-  });
-  it("Context should be set if setFallback === true", async () => {
-    const [, useI18nContext] = createChainedI18nContext({ dictionaries: dict, locale: "en" }, true);
-
-    const context = useI18nContext();
-
-    expect(context).not.toBe(null);
-  });
-  it("Locale switching works", async () => {
-    const [, useI18nContext] = createChainedI18nContext({ dictionaries: dict, locale: "en" }, true);
-
-    const [, { locale, setLocale }] = useI18nContext()!;
-
-    expect(locale()).toBe("en");
-
-    setLocale("fr");
-    expect(locale()).toBe("fr");
-
-    setLocale("en");
-    expect(locale()).toBe("en");
-  });
-  it("Translations work", async () => {
-    const [, useI18nContext] = createChainedI18nContext({ dictionaries: dict, locale: "en" }, true);
-
-    const [t, { setLocale }] = useI18nContext()!;
-
-    expect(t.hello({ name: "Tester" })).toBe("hello Tester, how are you?");
-    expect(t.goodbye({ name: "Tester" })).toBe("goodbye Tester");
-    expect(t.food.meat()).toBe("meat");
-
-    setLocale("fr");
-    expect(t.hello({ name: "Tester" })).toBe("bonjour Tester, comment vas-tu ?");
-    expect(t.goodbye({ name: "Tester" })).toBe("au revoir Tester");
-    expect(t.food.meat()).toBe("viande");
+    expect(flat).toEqual({
+      ...en_dict,
+      "numbers.1": "one",
+      "numbers.2": "two",
+      "numbers.3": "three",
+      "data.class": en_dict.data.class,
+      "data.currency": en_dict.data.currency,
+      "data.currency.name": "dollar",
+      "data.currency.symbol": "$",
+      "data.currency.iso": "USD",
+      "data.currency.to.usd": 1,
+      "data.users": en_dict.data.users,
+      "data.users.0": en_dict.data.users[0],
+      "data.users.0.name": "John",
+      "data.users.1": en_dict.data.users[1],
+      "data.users.1.name": "Kate",
+      "data.formatList": en_dict.data.formatList,
+    } satisfies typeof flat);
   });
 });
 
-describe("createChainedI18nDictionary", () => {
-  it("Translations work", async () => {
-    const dictionaries = createChainedI18nDictionary(dict);
-    const [locale, setLocale] = createSignal<keyof typeof dict>("en");
+describe("chainedResolver", () => {
+  test("initial", () => {
+    const flat_dict = i18n.flatten(en_dict);
+    const t = i18n.translator(() => flat_dict, i18n.resolveTemplate);
 
-    expect(dictionaries[locale()].hello({ name: "Tester" })).toBe("hello Tester, how are you?");
-    expect(dictionaries[locale()].goodbye({ name: "Tester" })).toBe("goodbye Tester");
-    expect(dictionaries[locale()].food.meat()).toBe("meat");
+    const chained = i18n.chained(en_dict, t);
+    const hello = chained.hello({ name: "Tester", thing: "day" });
+    expect(hello).toBe("Hello Tester! How is your day?");
 
-    setLocale("fr");
-    expect(dictionaries[locale()].hello({ name: "Tester" })).toBe(
-      "bonjour Tester, comment vas-tu ?",
-    );
-    expect(dictionaries[locale()].goodbye({ name: "Tester" })).toBe("au revoir Tester");
-    expect(dictionaries[locale()].food.meat()).toBe("viande");
+    const number1 = chained.numbers[1]();
+    expect(number1).toBe("one");
+
+    const data_class = chained.data.class();
+    expect(data_class).toBe(en_dict.data.class);
+
+    const currency_name = chained.data.currency.name();
+    expect(currency_name).toBe("dollar");
+
+    const currency_to_usd = chained.data.currency["to.usd"]();
+    expect(currency_to_usd).toBe(1);
+
+    const users = chained.data.users();
+    expect(users).toEqual(en_dict.data.users);
+
+    const format_list = chained.data.formatList(["John", "Kate", "Tester"]);
+    expect(format_list).toBe("John, Kate and Tester");
   });
 });
